refactor: move VirtualListHelper to helpers directory
